Crate roaring

This is a Rust port of the Roaring bitmap data structure, initially defined as a Java library and described in Better bitmap performance with Roaring bitmaps.

Re-exports

pub use bitmap::RoaringBitmap;

pub use treemap::RoaringTreemap;

Modules

bitmap

A compressed bitmap using the Roaring bitmap compression scheme.

treemap

A compressed bitmap with u64 values. Implemented as a BTreeMap of RoaringBitmaps.

Structs

NonSortedIntegers

An error type that is returned when an iterator isn’t sorted.

Traits

MultiOps

A Iterator::collect blanket implementation that provides extra methods for RoaringBitmap and RoaringTreemap.