Crate roaring

This is a Rust port of the Roaring bitmap data structure, initially defined as a Java library and described in Better bitmap performance with Roaring bitmaps.

Structs

DifferenceIter
An iterator for RoaringBitmap.

IntersectionIter
An iterator for RoaringBitmap.

Iter
An iterator for RoaringBitmap.

RoaringBitmap
A compressed bitmap using the Roaring bitmap compression scheme.

SymmetricDifferenceIter
An iterator for RoaringBitmap.

UnionIter
An iterator for RoaringBitmap.